Added checks to see if file exists

+++ ClassManager.py
@@ -5,29 +5,38 @@ now = datetime.datetime.now()
 def createClass():
     #Creates txt file for class roster
     className = input("Type name of class you wish to create:")
-    Class = open(className + " Roster", "w")
-    Class.close() #Closes the file
+    file = (className + " Roster")
+    if os.path.isfile(file):  #Checks if file exists
+        print("Class already exists.")
+    else:
+        Class = open(className + " Roster", "w")
+        Class.close() #Closes the file
 
 def addStudent():
     #Determines class file to access
     Choice = input("Which class would you like to add to?")
-    Class = open(Choice + " Roster", "r+")
+    
+    file = (Choice + " Roster")
+    if os.path.isfile(file):  #Checks if file exists
+        Class = open(file, "r+")
 
-    #Writes new student information in file
-    name = input("Type Student's Name:")
-    ID = input("Scan ID Card")
-    Class.writelines([name, ",", ID, ",\n"]) #\n makes program go to next line
-    while True:
-        repeatFunction = input("Would you like to add another student? [Y/N]")
-        if repeatFunction == "N":
-            break #Exits loop if "N" is given
-        elif repeatFunction == "Y":
-            name = input("Type Student's Name:")
-            ID = input("Scan ID Card")
-            Class.writelines([name, ",", ID, ";\n"])
-        else:
-            print("Invalid Input")
-    Class.close()
+        #Writes new student information in file
+        name = input("Type Student's Name:")
+        ID = input("Scan ID Card")
+        Class.writelines([name, ",", ID, ",\n"]) #\n makes program go to next line
+        while True:
+            repeatFunction = input("Would you like to add another student? [Y/N]")
+            if repeatFunction == "N":
+                break #Exits loop if "N" is given
+            elif repeatFunction == "Y":
+                name = input("Type Student's Name:")
+                ID = input("Scan ID Card")
+                Class.writelines([name, ",", ID, ";\n"])
+            else:
+                print("Invalid Input")
+        Class.close()
+    else:
+        print("Invalid Input")
 
 def rosterList(className):
     roster = open(className +" Roster", "r") #Opens Class Roster File in reading mode
@@ -40,25 +49,28 @@ def rosterList(className):
 def classCheckin():
    #Determines class list which will be searched
     Class = input("Which class do you want to take attendence for?")
-    
-    CheckIn = open(Class +" Roster", "r+")
-    lines = CheckIn.readlines()
-    Present = []
-    while True:
-        ID = input("Please Scan ID Card: ")
-        if ID == "exit":
-            break
-        else:
-            for row in lines:
-                if row.find(ID) != -1:
-                    presentName = row.rstrip(",1234567890 \n") #Strips ending ID off of text
-                    Present.insert(1, presentName)
-    CheckIn.close
+    file = (Class + " Roster")
+    if os.path.isfile(file):  #Checks if file exists
+        CheckIn = open(Class +" Roster", "r+")
+        lines = CheckIn.readlines()
+        Present = []
+        while True:
+            ID = input("Please Scan ID Card: ")
+            if ID == "exit":
+                break
+            else:
+                for row in lines:
+                    if row.find(ID) != -1:
+                        presentName = row.rstrip(",1234567890 \n") #Strips ending ID off of text
+                        Present.insert(1, presentName)
+        CheckIn.close
 
-    #Creates list of all absent students
-    absent = rosterList(Class) #Creates list of all students
-    for item in Present:
-        absent.remove(item) #Removes name from absent if on Present
+        #Creates list of all absent students
+        absent = rosterList(Class) #Creates list of all students
+        for item in Present:
+            absent.remove(item) #Removes name from absent if on Present
+    else:
+        print("Class does not exist.")
 
     #Creates txt file that store attendence information
     Attendence = open(Class +" Attendence", "w")
@@ -71,10 +83,14 @@ def classCheckin():
 
 
 def Display(AttendenceClass):
-    #Prints Attendence File of specified class
-    date_string = now.strftime("%m/%d/%Y")
-    print("Attendence for: " + AttendenceClass + " Date: " + date_string)
-    display = open(AttendenceClass +" Attendence", "r")
-    lines = display.readlines()
-    for row in lines:
-        print(row)
+    file = (AttendenceClass + " Roster")
+    if os.path.isfile(file):  #Checks if file exists
+        #Prints Attendence File of specified class
+        date_string = now.strftime("%m/%d/%Y")
+        print("Attendence for: " + AttendenceClass + " Date: " + date_string)
+        display = open(AttendenceClass +" Attendence", "r")
+        lines = display.readlines()
+        for row in lines:
+            print(row)
+    else:
+        print("Attendence file does not exist.")
